Split SelectionManager tests out of the selection monolith
Move the MockEditorWidget provider and the tests that construct a real
SelectionManager (provider-backed selection paths and the elevation
regression) into test_selection_manager.cpp. The remaining
test_selection_system.cpp keeps only pure SelectionState and
coordinate-conversion cases.

Drop the unused createMockObject helper and the now-unneeded includes,
and rename the misnamed mock hit-test case (it never touched
SelectionManager) to reflect what it actually exercises.
