+++ tests/CMakeLists.txt
@@ -12,6 +12,7 @@ add_executable(general_tests
     general/test_reader_factory.cpp
     general/test_selection_system.cpp
     general/test_selection_state.cpp
+    general/test_selection_manager.cpp
     general/test_tile_hex_conversion.cpp
     general/test_game_data_path_resolver.cpp
     general/test_ddraw_ini.cpp
